Lambda expression instead of reflect Method
Using reflect methods is a real mess. Normal users trying to use the
library almost need to participate in a lesson learning how it works. By
using an interface instead of these horrible solution makes it much
easier to use and understand. In addition to that users don't need to
create separate methods to make the functions work anymore. People are
now able to use anonymous interfaces or if preferred lambda expressions.